Changing how App Handels Validation and Classes
