+++ Class/ImageReader.py
