+++ Device/__init__.py
@@ -1,0 +1,2 @@
+from .cisco import Cisco
+from .mikrotik import Mikrotik
