+++ Device/cisco.py
@@ -1,6 +1,6 @@
-from .                    import Device
-from Global.Class.Auth    import Username , Password
-from Global.Class.Network import Hostname, IPv4 , Port
+from Class.device  import Device
+from Class.Auth    import Username , Password
+from Class.Network import Hostname, IPv4 , Port
 
 class Cisco(Device):
     def __init__(self,*,
