--- Global/Constant/__init__.py
@@ -1,4 +1,0 @@
-import os
-
-ROOT       = os.getcwd()
-CONFIGFILE = '.configfiles/config.ini'
