+++ Typing/hostname.py
@@ -1,0 +1,3 @@
+class Hostname : # not compleate
+    def __init__(self,Hostname) -> None:
+        self.Hostname = Hostname
