+++ Typing/ipv6.py
@@ -1,0 +1,3 @@
+class IPv6 : # Not Compleated
+    def __init__(self , IP : str) -> None :
+        pass
