+++ Typing/phone.py
@@ -1,0 +1,3 @@
+class Phone :
+    def __init__(self , Input : str) -> None:
+        self.Input = Input
